CDD-2696: Updated the cms.utils to render the time series / content
